@@ -4,8 +4,8 @@ version: 1.2.3
 homepage: "null"
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ">=2.17.0 <3.0.0"
+  flutter: ">=2.17.0"
 
 dependencies:
   flutter:
