@@ -1,7 +1,6 @@
 library tdtxle_inputs_flutter;
 
-
-export 'package:tdtxle_inputs_flutter/inputs/imagen_perfil/imagen_perfil.dart';
 export 'package:tdtxle_inputs_flutter/inputs/chip_imput.dart';
-export 'package:tdtxle_inputs_flutter/inputs/select%20imput/select_imput.dart';
-
+export 'package:tdtxle_inputs_flutter/inputs/imagen_perfil/imagen_perfil.dart';
+export 'package:tdtxle_inputs_flutter/inputs/select imput/select_imput.dart';
+// export 'package:tdtxle_inputs_flutter/inputs/select_input/select_input.dart';
